requirements.txt
flask
gunicorn
scikit-learn
nltk
PyPDF2
python-docx